+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Added a `json` feature to enable reading and writing JSON bodies in requests
+  and responses.
+
+- Added a `request` module with the useful utilities for constructing HTTP
+  requests.
+
 - A separate feature `util` has been removed, now this functionality is always
   available.